i created endpoint number 11 and 12

+++ src/main/java/com/example/amazon_clone/Model/MerchantStock.java
@@ -19,6 +19,6 @@ public class MerchantStock {
     @NotNull(message = "the stock can not be empty")
     //@Size(min = 10,message = "have to be more than 10 at start")
     @Min(value = 10, message = "have to be more than 10 at start")
-    private int stock;
+    private double stock;
 
 }
